test(job): cover fulfill chains and document deferred nested resolves
